@@ -1,8 +1,18 @@
 require 'rails/generators'
 
-module SssAvatar
+module IsellAvatar
   module Generators
     class InstallGenerator < ::Rails::Generators::Base
+
+      def add_assets_to_spree
+        # "Injecting into File" avoids the need to override layouts
+        # adding scripts or stylesheets tags lines.
+#        inject_into_file 'vendor/assets/javascripts/spree/frontend/all.js', "\n//= require isell/frontend\n", after: "spree/frontend", verbose: true
+        inject_into_file 'vendor/assets/stylesheets/spree/frontend/all.css', "\n *= require isell_avatar/frontend\n", after: "spree/frontend", verbose: true
+      end
+
+
+
 
 #      source_root SssAvatar::Engine.root.join('app', 'views', 'spree') #Needed by copy_views
     
